requirements.txt
feedparser
requests
python-dotenv
pandas
openpyxl
